+++ CodeWars/Mathematics/PerimeterOfSquaresInARectangle/PerimeterOfSquaresInARectangle.py
@@ -22,8 +22,31 @@ def fib_list_gen(n):
         yield b
 
 
+# Iterator
+class Fibonacci:
+    def __init__(self, n):
+        self.count = 0
+        self.end = n
+        self.curr = 0
+        self.next = 1
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        if self.count == self.end:
+            raise StopIteration()
+
+        new_next = self.curr + self.next
+        self.curr = self.next
+        self.next = new_next
+
+        self.count += 1
+        return self.curr
+
+
 def perimeter(n):
-    return sum(fib_list_gen(n+1)) * 4
+    return sum(Fibonacci(n+1)) * 4
 
 
 class TestExercise(unittest.TestCase):